WEB/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
boto3==1.35.81
botocore==1.35.81
certifi==2024.12.14
channels==4.2.0
charset-normalizer==3.4.0
cramjam==2.9.1
distro==1.9.0
Django==4.2.16
django-storages==1.14.4
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastparquet==2024.2.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
multidict==6.1.0
numpy==1.24.4
openai==1.57.4
orjson==3.10.12
packaging==24.2
pandas==2.0.3
propcache==0.2.0
psycopg==3.2.3
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
yarl==1.15.2
gunicorn
django-cors-headers
neo4j
drf_yasg
langchain-community